Fix open issues. Continue working
