+++ app/src/main/java/com/example/myapplication/GuestAdapter.java
@@ -3,16 +3,36 @@ package com.example.myapplication;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
+import org.jetbrains.annotations.NotNull;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class GuestAdapter  extends  RecyclerView.Adapter<GuestViewHolder> {
+public class GuestAdapter  extends  RecyclerView.Adapter<GuestAdapter.GuestViewHolder> {
 
     List<Guest> Guests;
+
+    public static class GuestViewHolder extends RecyclerView.ViewHolder {
+
+        public ImageView GuestImage;
+        public TextView Name;
+        public TextView Age;
+        public TextView Sex;
+
+        public GuestViewHolder(@NotNull View itemView) {
+            super(itemView);
+            GuestImage = itemView.findViewById(R.id.g_Image);
+            Name = itemView.findViewById(R.id.g_Name);
+            Age = itemView.findViewById(R.id.g_Age);
+            Sex = itemView.findViewById((R.id.g_Sex));
+        }
+    }
 
     public GuestAdapter() {
         super();
@@ -28,9 +48,8 @@ public class GuestAdapter  extends  RecyclerView.Adapter<GuestViewHolder> {
     @NonNull
     @Override
     public GuestViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.activity_main,parent,false);
-        GuestViewHolder viewHolder = new GuestViewHolder(view);
-        return viewHolder;
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.guests,parent,false);
+        return new GuestViewHolder(view);
     }
 
     @Override
@@ -38,10 +57,8 @@ public class GuestAdapter  extends  RecyclerView.Adapter<GuestViewHolder> {
         Guest guest = Guests.get(position);
         holder.GuestImage.setImageResource(guest.GuestImage);
         holder.Name.setText(guest.Name);
-        holder.Age.setText(guest.Age);
+        holder.Age.setText(Integer.toString(guest.Age));
         holder.Sex.setText(guest.Sex);
-
-
     }
 
     @Override
