+++ app/src/main/res/layout/activity_main.xml
@@ -23,12 +23,11 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:queryHint="Enter name"
-        app:queryTextSize="9sp" />
+        app:queryHint="Enter name" />
 
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_guest"
+        android:id="@+id/rvGuest"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="5dp"
