+++ app/src/main/res/layout/guests.xml
@@ -6,6 +6,7 @@
     android:layout_height="150dp">
 
     <androidx.cardview.widget.CardView
+        android:id="@+id/guestDetail"
         android:layout_width="0dp"
         android:layout_height="0dp"
         android:layout_marginStart="10dp"
