+++ app/src/main/res/values/ids.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <item name="rv_guest" type="id">rv_guest</item>
-    <item name="guestimage" type="id">g_Image</item>
-    <item name="name" type="id">g_Name</item>
-    <item name="age" type="id">g_Age</item>
-    <item name="sex" type="id">g_Sex</item>
+    <item name="rvGuest" type="id"/>
+    <item name="g_Image" type="id"/>
+    <item name="g_Name" type="id"/>
+    <item name="g_Age" type="id"/>
+    <item name="g_Sex" type="id"/>
 </resources>
